+++ db/migrate/20230330125747_migrate_project_regions.rb
@@ -1,0 +1,7 @@
+class MigrateProjectRegions < ActiveRecord::Migration[7.0]
+  def up
+    Project.all.each do |project|
+      ProjectRegionMigrator.new(project).migrate_up!
+    end
+  end
+end