+++ db/schema.rb
@@ -10,7 +10,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema[7.0].define(version: 2023_03_27_093354) do
+ActiveRecord::Schema[7.0].define(version: 2023_03_30_125747) do
   create_table "contacts", id: :uuid, default: -> { "newid()" }, force: :cascade do |t|
     t.uuid "project_id"
     t.string "name", null: false
@@ -284,6 +284,7 @@ ActiveRecord::Schema[7.0].define(version: 2023_03_27_093354) do
     t.date "provisional_conversion_date"
     t.boolean "directive_academy_order", default: false
     t.boolean "sponsor_trust_required", default: false
+    t.string "region"
     t.index ["assigned_to_id"], name: "index_projects_on_assigned_to_id"
     t.index ["caseworker_id"], name: "index_projects_on_caseworker_id"
     t.index ["regional_delivery_officer_id"], name: "index_projects_on_regional_delivery_officer_id"