+++ spec/factories/conversion/project_factory.rb
@@ -12,6 +12,7 @@ FactoryBot.define do
     assigned_to { association :user, :caseworker, email: "user.#{SecureRandom.uuid}@education.gov.uk" }
     directive_academy_order { false }
     sponsor_trust_required { false }
+    region { Project.regions["london"] }
     regional_delivery_officer { association :user, :regional_delivery_officer }
 
     trait :with_conditions do